Make changes according to feedback
